tweaked Market

+++ src/main/java/app/Market.java
@@ -20,7 +20,7 @@ public class Market{
     
     public void sell(double ProductPrice){
         Random r = new Random();
-        this.priceSell = r.nextDouble(13);
+        this.priceSell = r.nextInt(13);
         
         bud = bud+priceSell;
         
@@ -29,7 +29,7 @@ public class Market{
     public void buy(double ProductPrice){
         Random r = new Random();
         int choose = r.nextInt(21);
-        this.priceMaterial = r.nextDouble(20)*noise[choose];
+        this.priceMaterial = r.nextInt(20)*noise[choose];
         
         bud = bud-priceMaterial;
